Reorganize App assets
